Abinash_Rajarajan/publisher/requirements.txt
functions-framework
google-cloud-storage
google-cloud-pubsub
google-cloud-secret-manager
requests
pandas
